feat: iniciando implementação do middleware de erro

+++ src/models/GenericModel.ts
@@ -1,10 +1,10 @@
-import { Model as M } from 'mongoose';
+import { Model as M, Document, isValidObjectId } from 'mongoose';
 import { Model } from '../interfaces/ModelInterface';
 
 export default abstract class GenericModel<T> implements Model<T> {
-  protected _modelMongoose: M<T>;
+  protected _modelMongoose: M<T & Document>;
   
-  constructor(modelMongoose: M<T>) {
+  constructor(modelMongoose: M<T & Document>) {
     this._modelMongoose = modelMongoose;
   }
 
@@ -17,15 +17,18 @@ export default abstract class GenericModel<T> implements Model<T> {
   }
 
   async readOne(id: string): Promise<T | null> {
+    if (!isValidObjectId(id)) return null;
     return this._modelMongoose.findById(id);
   }
 
   async update(id: string, entity: T): Promise<T | null> {
+    if (!isValidObjectId(id)) return null;
     return this._modelMongoose
       .findOneAndUpdate({ _id: id }, entity, { returnOriginal: false });
   }
 
-  delete(id: string): Promise<T | null> {
-    return await this._modelMongoose.findByIdAndDelete({ _id: id });
+  async delete(id: string): Promise<T | null> {
+    if (!isValidObjectId(id)) return null;
+    return this._modelMongoose.findByIdAndDelete({ _id: id });
   }
 }
